Change excludeWhen allowed values, fixes #24
If excludeWhen is not set, no exclusions take place and all values are
returned as-is.  excludeWhen = null and excludeWhen not set can be
differentiated in JSON but not in TOML, as TOML does not support nil
values. This changes excludeWhen allowed values to 'false-like', 'none'
or a list of values.