BREAKING CHANGE: any and all return None if all values are None
Prior to this change, combinedType = any/all would return False if
all the values in the list were None. In most cases, datasets have
all values None when data is missing, and in that case, assuming
negative when data is missing is inconsistent with the mapping for
single fields.

In addition, for the clinical applications that are being used,
value mappings are better imagined as open-world (absence of a
statement does not imply falsity) rather than closed-world, where
the absence of a statement means it is false. As an example, for
medication mappings, it is possible we have missed out a medication
in the antibiotics list, so the absence does not imply that a
patient did not receive that category of medication, in this case,
antibiotics. In general, we are more certain that something happened
(patient was given a certain medication), assuming no entry errors,
rather than something did not (patient did not receive antibiotics).